Refactor split conversations from remote chat.
